Workspace系APIをBeUtl.Frameworkに移動
