+++ src/BeUtl.Framework/BeUtl.Framework.csproj
@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
   <ItemGroup>
     <ProjectReference Include="..\BeUtl.Core\BeUtl.Core.csproj" />
-    <ProjectReference Include="..\BeUtl.ProjectSystem\BeUtl.ProjectSystem.csproj" />
+    <ProjectReference Include="..\BeUtl.Graphics\BeUtl.Graphics.csproj" />
   </ItemGroup>
   <Import Project="$(RootDirectory)\build\props\Avalonia.props" />
   <Import Project="$(RootDirectory)\build\props\ReactiveProperty.props" />
