+++ src/BeUtl.Framework/IWorkspace.cs
@@ -1,0 +1,14 @@
+﻿using BeUtl.Collections;
+
+namespace BeUtl.Framework;
+
+public interface IWorkspace : ITopLevel, IDisposable, IStorable
+{
+    ICoreList<IWorkspaceItem> Items { get; }
+
+    IDictionary<string, string> Variables { get; }
+
+    Version AppVersion { get; }
+
+    Version MinAppVersion { get; }
+}
