+++ src/BeUtl.Framework/IWorkspaceItem.cs
@@ -1,0 +1,5 @@
+﻿namespace BeUtl.Framework;
+
+public interface IWorkspaceItem : IStorable, ILogicalElement, ICoreObject
+{
+}
