+++ src/BeUtl.Framework/Services/IProjectService.cs
@@ -1,20 +1,18 @@
-﻿using BeUtl.ProjectSystem;
-
-using Reactive.Bindings;
+﻿using Reactive.Bindings;
 
 namespace BeUtl.Framework.Services;
 
 public interface IProjectService
 {
-    IObservable<(Project? New, Project? Old)> ProjectObservable { get; }
+    IObservable<(IWorkspace? New, IWorkspace? Old)> ProjectObservable { get; }
 
-    IReactiveProperty<Project?> CurrentProject { get; }
+    IReactiveProperty<IWorkspace?> CurrentProject { get; }
 
     IReadOnlyReactiveProperty<bool> IsOpened { get; }
 
-    Project? OpenProject(string file);
+    IWorkspace? OpenProject(string file);
 
     void CloseProject();
 
-    Project? CreateProject(int width, int height, int framerate, int samplerate, string name, string location);
+    IWorkspace? CreateProject(int width, int height, int framerate, int samplerate, string name, string location);
 }
